Uneven Entity Spawn Fixes (#89)
Very often, monsters and items are standing on top of entities that are defined after them in the entity lump. If we set up the spawn position at map launch up in a linear order we immediately run into race conditions (see c1a0d in Half-Life, where a scientist spawns inside of a transparent grate brush).

So after the map is initialized - and every entity has been spawned and adjusted - we call a virtual member function on every entity class that spawned to ensure they can call functions like `droptofloor()` and accurately test against a copy of the world.

This should, in theory, fix most issues we have with entities not spawning in the right place.

Reviewed-on: https://code.idtech.space/vera/nuclide/pulls/89
Co-authored-by: Marco Cawthorne <marco@icculus.org>
Co-committed-by: Marco Cawthorne <marco@icculus.org>